Move `Spire.h` to top level.
The previous path (`Spire/Source/SpireLib/include/Spire.h`) made it really inconvenient to set up include paths for projects using Spire.

The new approach lets clients easily choose whether to do:

    #include <Spire.h>

or:

    #include <Spire/Spire.h>

based on how they set up their include path (and either way, the use setting up the include path only needs to know where they cloned the Spire repository, and not anything about its internal directory structure).
