+++ Spire.h
